Make a separate script for the pool
