+++ Assets/Mirror/Editor/Weaver/Weaver.cs
@@ -64,6 +64,7 @@ namespace Mirror.Weaver
         public static TypeReference NetworkReaderType;
 
         public static TypeReference NetworkWriterType;
+        public static TypeReference NetworkWriterPoolType;
 
         public static MethodReference NetworkWriterCtor;
         public static MethodReference NetworkReaderCtor;
@@ -295,6 +296,7 @@ namespace Mirror.Weaver
             NetworkReaderCtor = Resolvers.ResolveMethod(NetworkReaderDef, CurrentAssembly, ".ctor");
 
             NetworkWriterType = NetAssembly.MainModule.GetType("Mirror.NetworkWriter");
+            NetworkWriterPoolType = NetAssembly.MainModule.GetType("Mirror.NetworkWriterPool");
             TypeDefinition NetworkWriterDef  = NetworkWriterType.Resolve();
 
             NetworkWriterCtor = Resolvers.ResolveMethod(NetworkWriterDef, CurrentAssembly, ".ctor");
@@ -334,7 +336,7 @@ namespace Mirror.Weaver
             SyncDictionaryType = NetAssembly.MainModule.GetType("Mirror.SyncDictionary`2");
 
             NetworkBehaviourDirtyBitsReference = Resolvers.ResolveProperty(NetworkBehaviourType, CurrentAssembly, "syncVarDirtyBits");
-            GetPooledWriterMethod = Resolvers.ResolveMethod(NetworkWriterType, CurrentAssembly, "GetPooledWriter");
+            GetPooledWriterMethod = Resolvers.ResolveMethod(NetworkWriterPoolType, CurrentAssembly, "GetPooledWriter");
 
             ComponentType = UnityAssembly.MainModule.GetType("UnityEngine.Component");
             ClientSceneType = NetAssembly.MainModule.GetType("Mirror.ClientScene");
