+++ Assets/Mirror/Runtime/NetworkWriter.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Collections.Generic;
 using System.IO;
 using System.Text;
 using UnityEngine;
@@ -9,8 +8,6 @@ namespace Mirror
     // Binary stream Writer. Supports simple types, buffers, arrays, structs, and nested types
     public class NetworkWriter
     {
-        // reuse all writers, saves tons of memory allocations in hotpath
-        static readonly Stack<NetworkWriter> writerPool = new Stack<NetworkWriter>();
         // cache encoding instead of creating it with BinaryWriter each time
         // 1000 readers before:  1MB GC, 30ms
         // 1000 readers after: 0.8MB GC, 18ms
@@ -21,8 +18,8 @@ namespace Mirror
         // create writer immediately with it's own buffer so no one can mess with it and so that we can resize it.
         readonly BinaryWriter writer = new BinaryWriter(new MemoryStream(), encoding);
 
-        bool polled;
-        readonly bool reusable;
+        internal bool polled;
+        internal readonly bool reusable;
 
         // 'int' is the best type for .Position. 'short' is too small if we send >32kb which would result in negative .Position
         // -> converting long to int is fine until 2GB of data (MAX_INT), so we don't have to worry about overflows here
@@ -33,38 +30,10 @@ namespace Mirror
             this.reusable = reusable;
         }
 
-        public static NetworkWriter GetPooledWriter()
-        {
-            if (writerPool.Count != 0)
-            {
-                NetworkWriter writer = writerPool.Pop();
-                writer.polled = false;
-                // reset cached writer length and position
-                writer.SetLength(0);
-                return writer;
-            }
-
-            return new NetworkWriter(true);
-        }
-
-        public static void Recycle(NetworkWriter writer)
-        {
-            if (writer == null)
-            {
-                Debug.LogWarning("Recycling null writers is not allowed, please check your code!");
-                return;
-            }
-            if (writer.reusable && !writer.polled)
-            {
-                writer.polled = true;
-                writerPool.Push(writer);
-            }
-        }
-
         public byte[] Close()
         {
             byte[] data = ToArray();
-            Recycle(this);
+            NetworkWriterPool.Recycle(this);
             return data;
         }
