+++ Assets/Mirror/Runtime/NetworkWriterPool.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7c1f87b05942db74284500d93e118a0a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
